asas/training/adk_agent-development-kit-crash-course-main/requirements.txt
google-adk
yfinance==0.2.56
psutil==5.9.5
litellm
google-generativeai==0.8.5
python-dotenv
